requirements.txt
streamlit>=1.28.0
streamlit-folium>=0.15.0
folium>=0.15.0
rasterio>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
Pillow>=10.0.0
plotly>=5.18.0
openpyxl>=3.1.0
scipy>=1.11.0
